requirements.txt
fastapi~=0.111.0
pydantic~=2.7.3
starlette~=0.37.2
python-dotenv~=1.0.1
python-dateutil~=2.9.0.post0
passlib~=1.7.4
tortoise-orm~=0.21.3
python-jose~=3.3.0
fastapi-mail~=1.4.1
